Extracted ToMethodCall in a separated class
